chore: 게시글 관련 파일 src/features/community/post 폴더로 이동
