+++ src/features/community/post/components/PostList.tsx
@@ -1,7 +1,8 @@
-import { CATEGORY_TO_BOARD_ID } from '@/lib/community/constants';
-import React, { RefObject, useCallback } from 'react';
+import { CATEGORY_TO_BOARD_ID } from '@/src/features/community/shared/constants/constants';
+import React, { RefObject, useCallback, useMemo } from 'react';
 import { ActivityIndicator, FlatList, ListRenderItem } from 'react-native';
 import styled from 'styled-components/native';
+import { Empty, EmptyText, ErrorContainer, ErrorText, RetryButton, RetryButtonText } from '../../shared/styles/styles';
 import { useGetPosts } from '../hooks/useGetPosts';
 import { AllowedCategory, PostsListItem, SortParam } from '../types';
 import PostListCard from './PostListCard';
@@ -22,21 +23,54 @@ const PostList = ({ category, sort, scrollRef, openModal }: Props) => {
     [openModal],
   );
 
+  const listEmpty = useMemo(
+    () => (
+      <Empty>
+        <EmptyText>No bookmarked posts.</EmptyText>
+      </Empty>
+    ),
+    [],
+  );
+
+  const handleLoadMore = useCallback(() => {
+    if (hasNextPage && !isFetchingNextPage) {
+      fetchNextPage();
+    }
+  }, [hasNextPage, isFetchingNextPage, fetchNextPage]);
+
+  // 초기 로딩 중이면서 데이터가 없을 때만 로딩 표시
+  if (isLoading && !posts?.length) {
+    return (
+      <LoadingContainer>
+        <ActivityIndicator />
+      </LoadingContainer>
+    );
+  }
+
+  if (isError) {
+    return (
+      <ErrorContainer>
+        <ErrorText>An error occurred.</ErrorText>
+        <RetryButton onPress={() => refetch()}>
+          <RetryButtonText>Try Again</RetryButtonText>
+        </RetryButton>
+      </ErrorContainer>
+    );
+  }
+
   return (
     <FlatList
-      data={posts}
+      data={posts ?? []}
       ref={scrollRef}
       keyExtractor={(item: PostsListItem) => String(item.postId)}
       renderItem={renderPost}
       showsVerticalScrollIndicator={false}
-      onEndReachedThreshold={0.4}
-      onEndReached={() => {
-        if (hasNextPage && !isFetchingNextPage) {
-          fetchNextPage();
-        }
-      }}
-      refreshing={!isLoading && isRefetching}
+      ListEmptyComponent={listEmpty}
+      refreshing={isRefetching}
       onRefresh={() => refetch()}
+      onEndReachedThreshold={0.4}
+      initialNumToRender={10}
+      onEndReached={handleLoadMore}
       ListFooterComponent={
         isFetchingNextPage ? (
           <FooterLoading>
@@ -53,4 +87,9 @@ export default PostList;
 
 const FooterLoading = styled.View`
   padding: 16px 0;
+`;
+const LoadingContainer = styled.View`
+  flex: 1;
+  justify-content: center;
+  align-items: center;
 `;
