+++ src/features/community/post/hooks/useGetPostDetail.ts
@@ -1,4 +1,4 @@
-import { getPostDetail } from '@/src/features/community/apis/post';
+import { getPostDetail } from '@/src/features/community/post/apis/post';
 import { useQuery } from '@tanstack/react-query';
 import { PostDetail } from '../types';
