+++ src/features/community/post/hooks/usePostDetail.ts
